requirements.txt
requests==2.27.1
virtualenv==20.7.2
pipenv==2021.5.29
# environ==1.0
python-dotenv==0.21.0
mysql-connector-python==8.0.31